implement dynamic rendering for user form depending on create-mode or not
